% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjd_seasadj.R
\name{extract_sa}
\alias{extract_sa}
\title{Extract SA series from single or batch results}
\usage{
extract_sa(results, as_ts = FALSE, na_for_errors = TRUE)
}
\description{
Extract SA series from single or batch results
}
