% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjd_backcast.R
\name{print.rjd_backcast_batch}
\alias{print.rjd_backcast_batch}
\title{Print method for backcast batch results}
\usage{
\method{print}{rjd_backcast_batch}(x, ...)
}
\description{
Print method for backcast batch results
}
