% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjd_forecast.R
\name{print.rjd_forecast_batch}
\alias{print.rjd_forecast_batch}
\title{Print method for forecast batch results}
\usage{
\method{print}{rjd_forecast_batch}(x, ...)
}
\description{
Print method for forecast batch results
}
